Fix bugs in GraphPoly algorithms, add NUnit tests, clean up xUnit remnants

+++ src/IAFahim.Algebra.GraphPoly/Rook.cs
@@ -6,7 +6,7 @@ namespace IAFahim.Algebra.GraphPoly
     public static unsafe class Rook
     {
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static bool IsValidPlacement(int n, int m, bool* blocked, int mask, bool* rowUsed, bool* colUsed, out int count)
+        private static bool IsValidPlacement(int n, int m, bool* blocked, long mask, bool* rowUsed, bool* colUsed, out int count)
         {
             count = 0;
             for (int i = 0; i < n; i++) rowUsed[i] = false;
@@ -15,8 +15,8 @@ namespace IAFahim.Algebra.GraphPoly
             for (int i = 0; i < n; i++)
             for (int j = 0; j < m; j++)
             {
-                long bit = (long)i * (long)m + (long)j;
-                if ((mask & (1 << (int)bit)) == 0) continue;
+                int bit = i * m + j;
+                if ((mask & (1L << bit)) == 0) continue;
                 if (blocked[bit] || rowUsed[i] || colUsed[j]) return false;
                 rowUsed[i] = true;
                 colUsed[j] = true;
@@ -29,10 +29,10 @@ namespace IAFahim.Algebra.GraphPoly
         public static long Run(int n, int m, bool* blocked, long x, int MOD)
         {
             long result = 0L;
-            int size = 1 << (n * m);
+            long size = 1L << (n * m);
             bool* rowUsed = stackalloc bool[n];
             bool* colUsed = stackalloc bool[m];
-            for (int mask = 0; mask < size; mask++)
+            for (long mask = 0; mask < size; mask++)
             {
                 if (IsValidPlacement(n, m, blocked, mask, rowUsed, colUsed, out int count))
                 {
